:16 Delete Account-2

+++ server/controllers/UsersController.js
@@ -1,4 +1,6 @@
 const User = require('../models/user')
+const Post = require('../models/post')
+const Collection = require('../models/collection')
 const fs = require('fs')
 
 module.exports = {
@@ -128,6 +130,10 @@ module.exports = {
 
     const deleteOldImageFile = user => {
       return new Promise((resolve, reject) => {
+        if (user.image == 'face-default.png') {
+          resolve(user)
+          return
+        }
         fs.unlink('./server/uploads/profile-images/' + user.image, function(
           err
         ) {
@@ -185,10 +191,61 @@ module.exports = {
   deleteAccount: (req, res) => {
     const { email } = req.params
     const { password } = req.body
+    var account = null
 
     const verify = user => {
+      account = user
       if (!user) throw new Error('User does not exist')
       else if (user.verify(password)) return User.deleteOneByEmail(email)
+    }
+
+    const deleteOldImageFile = user => {
+      return new Promise((resolve, reject) => {
+        if (user.image == 'face-default.png') {
+          resolve(user)
+          return
+        }
+        fs.unlink('./server/uploads/profile-images/' + account.image, function(
+          err
+        ) {
+          if (err) {
+            console.log(err)
+            reject(err)
+            return
+          }
+          resolve(user)
+        })
+      })
+    }
+
+    const deleteFiles = () => {
+      Post.findByAuthorId(account._id).then(posts => {
+        posts.forEach(post => {
+          fs.unlink('./server/uploads/posts/' + post.data.filename, err => {
+            if (err) {
+              console.log(err)
+              return Promise.reject(err)
+            }
+          })
+        })
+        return Promise.resolve()
+      })
+    }
+
+    const deletePosts = () => {
+      return Post.deleteByAuthorId(account._id)
+    }
+
+    const subtractFF = () => {
+      return User.subtractFF(account.followings, account.followers, account._id)
+    }
+
+    const deleteCollections = () => {
+      return Collection.deleteCollections(account.collections)
+    }
+
+    const decreaseLikes = () => {
+      return Post.decLikes(account.likes)
     }
 
     const respond = () => {
@@ -200,8 +257,19 @@ module.exports = {
     }
 
     User.findOneByEmail(email)
-    //TODO: 계정 제거 후 게시물, 컬렉션, 팔로우 등 삭제하기.
       .then(verify)
+      // 프로필 이미지 파일 삭제
+      .then(deleteOldImageFile)
+      // account.posts에 해당하는 post 삭제, post.data.filename 파일 삭제
+      .then(deleteFiles)
+      .then(deletePosts)
+      // account.followings에 해당하는 user의 followers에서 _id 를삭제
+      // account.followers 해당하는 user의 followings에 _id 를삭제
+      .then(subtractFF)
+      // collections 에 해당하는 collection 삭제
+      .then(deleteCollections)
+      // likes 에 해당하는 post.likes 1 감소
+      .then(decreaseLikes)
       .then(respond)
       .catch(onError)
   }
